# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SECOM MES Database API",
    "description": "REST API for SECOM Manufacturing Execution System Database\n\nThis API provides access to semiconductor manufacturing data including:\n- Master data (Equipment, Shifts, Operators, Product Types, Features)\n- Production data (Lots, Measurements, Quality Results)\n- Analytics (Production summaries, Equipment health, Shift performance, Quality analytics)\n\n**Database**: MariaDB 11.2\n**Total Records**: 1,567 lots with 590 sensor measurements each\n**Time Period**: July 2008 - September 2008\n\n**Key Endpoints**:\n- `/equipment` - Manufacturing equipment management\n- `/lots` - Production lot tracking with filters\n- `/measurements` - Sensor measurement data\n- `/quality` - Quality inspection results\n- `/analytics` - Analytics and reporting endpoints\n",
    "contact": {
      "name": "SECOM MES Team",
      "email": "mes@secom.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
    "/shifts/{shiftId}": {
      "get": {
        "tags": [
          "Shifts"
        ],
        "summary": "Get shift by ID",
        "description": "Returns a single shift by shift ID",
        "operationId": "getShiftById",
        "parameters": [
          {
            "name": "shiftId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Shift"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Shifts"
        ],
        "summary": "Update shift",
        "description": "Updates an existing shift record",
        "operationId": "updateShift",
        "parameters": [
          {
            "name": "shiftId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Shift"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Shift"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Shifts"
        ],
        "summary": "Delete shift",
        "description": "Deletes a shift record",
        "operationId": "deleteShift",
        "parameters": [
          {
            "name": "shiftId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/quality/results/{qualityResultId}": {
      "get": {
        "tags": [
          "Quality"
        ],
        "summary": "Get quality result by ID",
        "description": "Returns a single quality result by quality result ID",
        "operationId": "getResultById",
        "parameters": [
          {
            "name": "qualityResultId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/QualityResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Quality"
        ],
        "summary": "Update quality result",
        "description": "Updates an existing quality result",
        "operationId": "updateResult",
        "parameters": [
          {
            "name": "qualityResultId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityResult"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/QualityResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Quality"
        ],
        "summary": "Delete quality result",
        "description": "Deletes a quality result record",
        "operationId": "deleteResult",
        "parameters": [
          {
            "name": "qualityResultId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/shifts": {
      "get": {
        "tags": [
          "Shifts"
        ],
        "summary": "Get all shifts",
        "description": "Returns list of all work shifts",
        "operationId": "getAllShifts",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shift"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Shifts"
        ],
        "summary": "Create new shift",
        "description": "Creates a new shift record",
        "operationId": "createShift",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Shift"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Shift"
                }
              }
            }
          }
        }
      }
    },
    "/quality/results": {
      "get": {
        "tags": [
          "Quality"
        ],
        "summary": "Get all quality results",
        "description": "Returns paginated list of all quality inspection results",
        "operationId": "getAllResults",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageQualityResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Quality"
        ],
        "summary": "Create new quality result",
        "description": "Creates a new quality inspection result",
        "operationId": "createResult",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityResult"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/QualityResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/shifts/code/{code}": {
      "get": {
        "tags": [
          "Shifts"
        ],
        "summary": "Get shift by code",
        "description": "Returns shift by unique shift code (DAY, SWING, NIGHT)",
        "operationId": "getShiftByCode",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Shift"
                }
              }
            }
          }
        }
      }
    },
    "/quality/results/passed": {
      "get": {
        "tags": [
          "Quality"
        ],
        "summary": "Get passed lots",
        "description": "Returns all lots that passed quality inspection (classification = -1)",
        "operationId": "getPassedResults",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageQualityResult"
                }
              }
            }
          }
        }
      }
    },
    "/quality/results/lot/{lotId}": {
      "get": {
        "tags": [
          "Quality"
        ],
        "summary": "Get quality result for a lot",
        "description": "Returns quality result for a specific lot",
        "operationId": "getResultByLotId",
        "parameters": [
          {
            "name": "lotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/QualityResult"
                }
              }
            }
          }
        }
      }
    },
    "/quality/results/high-risk": {
      "get": {
        "tags": [
          "Quality"
        ],
        "summary": "Get high-risk lots",
        "description": "Returns lots with predicted risk above threshold (default 0.7)",
        "operationId": "getHighRiskResults",
        "parameters": [
          {
            "name": "threshold",
            "in": "query",
            "description": "Risk threshold (0.0 to 1.0)",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.7
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QualityResult"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/quality/results/failed": {
      "get": {
        "tags": [
          "Quality"
        ],
        "summary": "Get failed lots",
        "description": "Returns all lots that failed quality inspection (classification = 1)",
        "operationId": "getFailedResults",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageQualityResult"
                }
              }
            }
          }
        }
      }
    },
    "/quality/results/defect/{type}": {
      "get": {
        "tags": [
          "Quality"
        ],
        "summary": "Get results by defect type",
        "description": "Returns quality results filtered by defect type (electrical_fail, dimensional_oor, etc.)",
        "operationId": "getResultsByDefectType",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QualityResult"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "LocalTime": {
        "type": "object",
        "properties": {
          "hour": {
            "type": "integer",
            "format": "int32"
          },
          "minute": {
            "type": "integer",
            "format": "int32"
          },
          "second": {
            "type": "integer",
            "format": "int32"
          },
          "nano": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Shift": {
        "type": "object",
        "properties": {
          "shiftId": {
            "type": "integer",
            "format": "int32"
          },
          "shiftCode": {
            "type": "string"
          },
          "shiftName": {
            "type": "string"
          },
          "startTime": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "endTime": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Equipment": {
        "type": "object",
        "properties": {
          "equipmentId": {
            "type": "integer",
            "format": "int32"
          },
          "equipmentCode": {
            "type": "string"
          },
          "equipmentName": {
            "type": "string"
          },
          "equipmentType": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "manufacturer": {
            "type": "string"
          },
          "installDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Lot": {
        "type": "object",
        "properties": {
          "lotId": {
            "type": "integer",
            "format": "int32"
          },
          "lotNumber": {
            "type": "string"
          },
          "productType": {
            "$ref": "#/components/schemas/ProductType"
          },
          "equipment": {
            "$ref": "#/components/schemas/Equipment"
          },
          "operator": {
            "$ref": "#/components/schemas/Operator"
          },
          "shift": {
            "$ref": "#/components/schemas/Shift"
          },
          "productionStart": {
            "type": "string",
            "format": "date-time"
          },
          "productionEnd": {
            "type": "string",
            "format": "date-time"
          },
          "waferCount": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Operator": {
        "type": "object",
        "properties": {
          "operatorId": {
            "type": "integer",
            "format": "int32"
          },
          "operatorCode": {
            "type": "string"
          },
          "operatorName": {
            "type": "string"
          },
          "employeeNumber": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "hireDate": {
            "type": "string",
            "format": "date"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProductType": {
        "type": "object",
        "properties": {
          "productTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "productCode": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "productFamily": {
            "type": "string"
          },
          "targetYield": {
            "type": "number"
          },
          "specificationVersion": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "QualityResult": {
        "type": "object",
        "properties": {
          "resultId": {
            "type": "integer",
            "format": "int32"
          },
          "lot": {
            "$ref": "#/components/schemas/Lot"
          },
          "classification": {
            "type": "integer",
            "format": "int32"
          },
          "testTimestampRaw": {
            "type": "string"
          },
          "testDatetime": {
            "type": "string",
            "format": "date-time"
          },
          "predictedRisk": {
            "type": "number"
          },
          "riskScore": {
            "type": "number"
          },
          "riskFactors": {
            "type": "string"
          },
          "modelVersion": {
            "type": "string"
          },
          "qualityScore": {
            "type": "number"
          },
          "defectType": {
            "type": "string"
          },
          "defectCode": {
            "type": "string"
          },
          "defectLocation": {
            "type": "string"
          },
          "inspector": {
            "$ref": "#/components/schemas/Operator"
          },
          "notes": {
            "type": "string"
          },
          "reviewer": {
            "$ref": "#/components/schemas/Operator"
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time"
          },
          "disposition": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PageQualityResult": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QualityResult"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "last": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "paged": {
            "type": "boolean"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}